requirements.txt
GitPython
httpx[http2]
keyring
keyrings.alt # make optional
pillow
psutil
